Buffs and fixes
Tumblebug doesn't push sides when tossing a pawn anymore.
Tumblebug takes into account having nowhere to toss a pawn when checking for TwoClickException.
Moth can now reposition at a minimum range of 2 instead of 3 if it's not pushing anything to that tile.
Digger's former upgrades are merged. A new upgrade lets it encase a targeted building in stone, replacing it with a 1 HP rock - basically a themed shield.